Rename Int->Int64 (#255)
* Rename Int->Int64

* Update readme & changelog

+++ tests/test_benchmark_1d.py
@@ -5,7 +5,7 @@ import boost_histogram as bh
 import numpy as np
 from numpy.testing import assert_array_equal, assert_allclose
 
-STORAGES = (bh.storage.Int, bh.storage.Double, bh.storage.Unlimited)
+STORAGES = (bh.storage.Int64, bh.storage.Double, bh.storage.Unlimited)
 DTYPES = (np.float64, np.float32, np.int64, np.int32)
 
 bins = 100
